requirements.txt
numpy>=1.17.2
scipy>=1.3.1
matplotlib>=3.1.1
pandas>=0.25.1
joblib>=0.13.2 
scikit-learn>=0.21.3
Pillow>=6.1.0
imageio>=2.5.0
rasterio==1.0.14
jupyter>=1.0.0 